refactor: organize package exported types
