+++ src/geometries/boundingBox.ts
@@ -3,7 +3,7 @@ import type { TileMatrixSet } from '../tiles/tileMatrixSet';
 import { TileRange } from '../tiles/tileRange';
 import { avoidNegativeZero, clampValues, tileEffectiveHeight, tileEffectiveWidth } from '../tiles/tiles';
 import type { TileMatrixId } from '../tiles/types';
-import type { ArrayElement } from '../types';
+import type { ArrayElement } from '../utils/types';
 import { validateBoundingBoxByTileMatrix, validateCRSByOtherCRS, validateMetatile } from '../validations/validations';
 import { Point } from './point';
 import { Polygon } from './polygon';
