+++ src/geometries/geometry.ts
@@ -5,9 +5,9 @@ import { Tile } from '../tiles/tile';
 import type { TileMatrixSet } from '../tiles/tileMatrixSet';
 import { positionToTileIndex, tileMatrixToBBox } from '../tiles/tiles';
 import type { CRS as CRSType } from '../tiles/types';
-import type { ArrayElement, CoordRefSysJSON } from '../types';
+import type { ArrayElement } from '../utils/types';
 import { validateCRS, validateCRSByOtherCRS, validateMetatile } from '../validations/validations';
-import type { GeoJSONBaseGeometry, GeoJSONGeometry, JSONFGFeature } from './types';
+import type { CoordRefSysJSON, GeoJSONBaseGeometry, GeoJSONGeometry, JSONFGFeature } from './types';
 
 /**
  * Geometry class
