+++ src/tiles/types.ts
@@ -1,5 +1,7 @@
-import { ArrayElement } from '../types';
+import { ArrayElement } from '../utils/types';
 import { CORNER_OF_ORIGIN_CODE } from './constants';
+
+export type Comparison = 'equal' | 'closest' | 'lower' | 'higher';
 
 /**
  * Tile index
