+++ src/types.ts
@@ -1,17 +1,8 @@
-import type { TileMatrixSet, TileMatrixSetJSON } from './tiles/types';
-
-type NonOptional<Type> = {
-  [Property in keyof Type]-?: Type[Property];
-};
-
-export type ArrayElement<T> = T extends (infer U)[] ? U : never;
-
-export type Comparison = 'equal' | 'closest' | 'lower' | 'higher';
-export interface CoordRefSys {
-  coordRefSys?: TileMatrixSet['crs']; // TODO: change type according to - OGC Features and Geometries JSON - Part 1: Core
-}
-export interface CoordRefSysJSON {
-  coordRefSys?: TileMatrixSetJSON['crs']; // TODO: change type according to - OGC Features and Geometries JSON - Part 1: Core
-}
-export type ConcreteCoordRefSys = NonOptional<CoordRefSys>;
-export type ConcreteCoordRefSysJSON = NonOptional<CoordRefSysJSON>;
+export { BoundingBox } from './geometries/boundingBox';
+export { GeometryCollection } from './geometries/geometryCollection';
+export { Line } from './geometries/line';
+export { Point } from './geometries/point';
+export { Polygon } from './geometries/polygon';
+export { area, distance, geodesicDistance, geometriesEqual } from './measurements/measurements';
+export { Tile } from './tiles/tile';
+export { TileMatrixSet } from './tiles/tileMatrixSet';
