+++ src/utils/types.ts
@@ -1,0 +1,1 @@
+export type ArrayElement<T> = T extends (infer U)[] ? U : never;
